refactor(file-structure): split tutorial graphql inputs and types
