+++ apps/project/graphql/types/project_types/base.py
@@ -1,0 +1,67 @@
+import strawberry
+
+from utils.geo.tile_server.config import TileServerNameEnum
+from utils.geo.tile_server.models import (
+    TileServerCommonConfig,
+    TileServerConfig,
+    TileServerCustomConfig,
+)
+from utils.geo.vector_tile_server.config import VectorTileServerNameEnum
+from utils.geo.vector_tile_server.models import (
+    VectorTileServerCommonConfig,
+    VectorTileServerConfig,
+    VectorTileServerCustomConfig,
+)
+
+
+@strawberry.type
+class VectorTileServerType:
+    type: VectorTileServerNameEnum
+    label: str
+    url: str
+    layers: list[str]
+    min_zoom: int | None
+    max_zoom: int | None
+    credits: str
+
+
+@strawberry.type
+class RasterTileServerType:
+    type: TileServerNameEnum
+    label: str
+    url: str
+    # FIXME(tnagorra): We can implement min_zoom and max_zoom
+    # after we update rendering of tiles in the app from images to maplibre
+    credits: str
+
+
+@strawberry.type
+class TileServersType:
+    vector: list[VectorTileServerType]
+    raster: list[RasterTileServerType]
+
+
+# Tile server
+@strawberry.experimental.pydantic.type(model=TileServerCustomConfig, all_fields=True)
+class ProjectTileServerCustomConfig: ...
+
+
+@strawberry.experimental.pydantic.type(model=TileServerCommonConfig, all_fields=True)
+class ProjectTileServerCommonConfig: ...
+
+
+@strawberry.experimental.pydantic.type(model=TileServerConfig, all_fields=True)
+class ProjectTileServerConfig: ...
+
+
+# Vector tile server
+@strawberry.experimental.pydantic.type(model=VectorTileServerCustomConfig, all_fields=True)
+class ProjectVectorTileServerCustomConfig: ...
+
+
+@strawberry.experimental.pydantic.type(model=VectorTileServerCommonConfig, all_fields=True)
+class ProjectVectorTileServerCommonConfig: ...
+
+
+@strawberry.experimental.pydantic.type(model=VectorTileServerConfig, all_fields=True)
+class ProjectVectorTileServerConfig: ...
