+++ apps/tutorial/graphql/inputs/project_types/completeness.py
@@ -1,0 +1,7 @@
+import strawberry
+
+from project_types.tile_map_service.completeness import tutorial as completeness_tutorial
+
+
+@strawberry.experimental.pydantic.input(model=completeness_tutorial.CompletenessTutorialTaskProperty, all_fields=True)
+class CompletenessTutorialTaskPropertyInput: ...
