+++ apps/tutorial/graphql/types/project_types/compare.py
@@ -1,0 +1,7 @@
+import strawberry
+
+from project_types.tile_map_service.compare import tutorial as compare_tutorial
+
+
+@strawberry.experimental.pydantic.type(model=compare_tutorial.CompareTutorialTaskProperty, all_fields=True)
+class CompareTutorialTaskPropertyType: ...
