+++ apps/tutorial/graphql/types/project_types/validate.py
@@ -1,0 +1,7 @@
+import strawberry
+
+from project_types.validate import tutorial as validate_tutorial
+
+
+@strawberry.experimental.pydantic.type(model=validate_tutorial.ValidateTutorialTaskProperty, all_fields=True)
+class ValidateTutorialTaskPropertyType: ...
